# Conflicts:
#	packages/kbn-optimizer/limits.yml
#	src/plugins/chart_expressions/expression_xy/common/__mocks__/index.ts
#	src/plugins/chart_expressions/expression_xy/common/constants.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/annotation_layer.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/common_data_layer_args.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/common_xy_args.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/data_layer.test.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/extended_annotation_layer.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/extended_data_layer.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/index.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/layered_xy_vis.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/legend_config.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/reference_line_layer.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/xy_vis.test.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/xy_vis.ts
#	src/plugins/chart_expressions/expression_xy/common/expression_functions/xy_vis_fn.ts
#	src/plugins/chart_expressions/expression_xy/common/helpers/index.ts
#	src/plugins/chart_expressions/expression_xy/common/helpers/layers.test.ts
#	src/plugins/chart_expressions/expression_xy/common/helpers/layers.ts
#	src/plugins/chart_expressions/expression_xy/common/i18n/index.tsx
#	src/plugins/chart_expressions/expression_xy/common/index.ts
#	src/plugins/chart_expressions/expression_xy/common/types/expression_functions.ts
#	src/plugins/chart_expressions/expression_xy/common/types/expression_renderers.ts
#	src/plugins/chart_expressions/expression_xy/common/utils/log_datatables.ts
#	src/plugins/chart_expressions/expression_xy/public/__mocks__/index.tsx
#	src/plugins/chart_expressions/expression_xy/public/components/__snapshots__/xy_chart.test.tsx.snap
#	src/plugins/chart_expressions/expression_xy/public/components/annotations.tsx
#	src/plugins/chart_expressions/expression_xy/public/components/data_layers.tsx
#	src/plugins/chart_expressions/expression_xy/public/components/legend_action.test.tsx
#	src/plugins/chart_expressions/expression_xy/public/components/legend_action.tsx
#	src/plugins/chart_expressions/expression_xy/public/components/reference_lines.test.tsx
#	src/plugins/chart_expressions/expression_xy/public/components/reference_lines.tsx
#	src/plugins/chart_expressions/expression_xy/public/components/x_domain.tsx
#	src/plugins/chart_expressions/expression_xy/public/components/xy_chart.test.tsx
#	src/plugins/chart_expressions/expression_xy/public/components/xy_chart.tsx
#	src/plugins/chart_expressions/expression_xy/public/helpers/axes_configuration.ts
#	src/plugins/chart_expressions/expression_xy/public/helpers/color_assignment.ts
#	src/plugins/chart_expressions/expression_xy/public/helpers/data_layers.tsx
#	src/plugins/chart_expressions/expression_xy/public/helpers/index.ts
#	src/plugins/chart_expressions/expression_xy/public/helpers/interval.ts
#	src/plugins/chart_expressions/expression_xy/public/helpers/layers.ts
#	src/plugins/chart_expressions/expression_xy/public/helpers/state.ts
#	src/plugins/chart_expressions/expression_xy/public/helpers/visualization.ts
#	src/plugins/chart_expressions/expression_xy/public/plugin.ts
#	src/plugins/chart_expressions/expression_xy/server/plugin.ts
#	src/plugins/event_annotation/common/index.ts
#	src/plugins/event_annotation/common/manual_event_annotation/index.ts
#	src/plugins/event_annotation/common/types.ts
#	src/plugins/expressions/common/execution/execution.test.ts
#	src/plugins/expressions/common/execution/execution.ts
#	x-pack/plugins/lens/public/editor_frame_service/editor_frame/expression_helpers.ts
#	x-pack/plugins/lens/public/editor_frame_service/editor_frame/suggestion_panel.tsx
#	x-pack/plugins/lens/public/shared_components/legend_settings_popover.tsx
#	x-pack/plugins/lens/public/xy_visualization/__snapshots__/to_expression.test.ts.snap
#	x-pack/plugins/lens/public/xy_visualization/to_expression.ts
#	x-pack/plugins/lens/public/xy_visualization/types.ts
#	x-pack/plugins/lens/public/xy_visualization/visualization.tsx
#	x-pack/plugins/lens/public/xy_visualization/xy_config_panel/annotations_config_panel/annotations_panel.tsx
#	x-pack/plugins/lens/public/xy_visualization/xy_config_panel/color_picker.tsx
#	x-pack/plugins/lens/public/xy_visualization/xy_config_panel/reference_line_config_panel/reference_line_panel.tsx
#	x-pack/plugins/lens/public/xy_visualization/xy_config_panel/shared/icon_select.tsx
#	x-pack/plugins/lens/public/xy_visualization/xy_config_panel/shared/marker_decoration_settings.tsx
#	x-pack/plugins/lens/server/embeddable/make_lens_embeddable_factory.ts
#	x-pack/plugins/lens/server/migrations/common_migrations.ts
#	x-pack/plugins/lens/server/migrations/saved_object_migrations.test.ts
#	x-pack/plugins/lens/server/migrations/saved_object_migrations.ts

+++ x-pack/plugins/osquery/server/routes/saved_query/delete_saved_query_route.ts
@@ -9,8 +9,10 @@ import { schema } from '@kbn/config-schema';
 import { IRouter } from '@kbn/core/server';
 import { PLUGIN_ID } from '../../../common';
 import { savedQuerySavedObjectType } from '../../../common/types';
+import { OsqueryAppContext } from '../../lib/osquery_app_context_services';
+import { isSavedQueryPrebuilt } from './utils';
 
-export const deleteSavedQueryRoute = (router: IRouter) => {
+export const deleteSavedQueryRoute = (router: IRouter, osqueryContext: OsqueryAppContext) => {
   router.delete(
     {
       path: '/internal/osquery/saved_query/{id}',
@@ -22,7 +24,13 @@ export const deleteSavedQueryRoute = (router: IRouter) => {
       options: { tags: [`access:${PLUGIN_ID}-writeSavedQueries`] },
     },
     async (context, request, response) => {
-      const savedObjectsClient = context.core.savedObjects.client;
+      const coreContext = await context.core;
+      const savedObjectsClient = coreContext.savedObjects.client;
+
+      const isPrebuilt = await isSavedQueryPrebuilt(osqueryContext, request.params.id);
+      if (isPrebuilt) {
+        return response.conflict({ body: `Elastic prebuilt Saved query cannot be deleted.` });
+      }
 
       await savedObjectsClient.delete(savedQuerySavedObjectType, request.params.id, {
         refresh: 'wait_for',
